+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Read Event Log simulation high scores from the SnakeLab database instead of the Ax3l database.
+
 ## [1.3.8] - 2026-09-20 @ 06:40
 
 ### Changed